server/requirements.txt
django
psycopg
gunicorn
django-configurations
djangorestframework
drf-yasg
django-cors-headers
django-filter
requests
celery
django-celery-results
redis
